Example services: mount /dev/pts and /dev/shm in early-filesystems
Also, don't try and clear /var/lock and /var/run, we should assume they
are symlinks back to /run[/lock] which doesn't survive reboot anyway.

+++ doc/linux/services/rcboot.sh
@@ -9,15 +9,15 @@ if [ "$1" != "stop" ]; then
   # cleanup
   # (delete /tmp etc)
   rm -rf /tmp/* /tmp/.[!.]* /tmp/..?*
-  rm -rf /var/lock/* /var/lock/.[!.]* /var/lock/..?*
-  rm -rf /var/run/* /var/run/.[!.]* /var/run/..?*
-  # create state directories
+  
+  # empty utmp, create needed directories
   : > /var/run/utmp
-  mkdir /var/run/dbus
-  chmod og+rx /var/run/dbus
+  mkdir -m og-w /var/run/dbus
 
   # Configure random number generator
-  cat /var/state/random-seed > /dev/urandom
+  if [ -e /var/state/random-seed ]; then
+    cat /var/state/random-seed > /dev/urandom;
+  fi
   
   # Configure network
   /sbin/ifconfig lo 127.0.0.1
